Parallel streaming phase 1 + phase-2 data structures (milestone 4)
Phase 1: files stream in 32MB whitespace-bounded chunks to worker
threads (all cores), counted into thread-local FxHash maps, merged
first-seen. Never holds a whole file in memory. Phase 2: vendored
FxHash replaces SipHash; per-pair position HashSets become Vec<u32>
with adjacent dedup (semantics identical, ~5x smaller); symbols shrink
to 8 bytes. Parity CI green; 1GB FineWeb: 39.3s -> 15.3s (3.6x vs HF
on 10 cores), RSS 3.4GB -> 2.4GB.

+++ gigatrain/src/word.rs
@@ -11,7 +11,7 @@ pub type Pair = (u32, u32);
 pub struct Symbol {
     pub c: u32,
     /// Length in chars (HF counts 1 per char; merged symbols sum).
-    pub len: usize,
+    pub len: u32,
 }
 
 #[derive(Clone, Debug, Default)]
@@ -55,14 +55,14 @@ impl Word {
                 if w > 0 {
                     let prev = syms[w - 1];
                     changes.push(((prev.c, c1), -1));
-                    if prev.len + new_len < max_len {
+                    if (prev.len as usize + new_len as usize) < max_len {
                         changes.push(((prev.c, new_id), 1));
                     }
                 }
                 if k + 2 < n {
                     let right = syms[k + 2];
                     changes.push(((c2, right.c), -1));
-                    if right.len + new_len < max_len {
+                    if (right.len as usize + new_len as usize) < max_len {
                         changes.push(((new_id, right.c), 1));
                     }
                 }
